+++ src/main/java/truenorth/devops/lab/VHS/repository/UserRepository.java
@@ -9,7 +9,7 @@ import truenorth.devops.lab.VHS.model.User;
 public interface UserRepository extends JpaRepository<User, Long>{
 
 	@Query(value="SELECT * FROM korisnik WHERE user_id = :id", nativeQuery=true)
-	public User getUserById(@Param("id") int id);
+	public User getUserById(@Param("id") long id);
 	
 	@Query(value="SELECT * FROM korisnik WHERE username = :username", nativeQuery=true)
 	public User getUserByUsername(@Param("username") String username);
