+++ src/main/java/truenorth/devops/lab/VHS/repository/VHSRepository.java
@@ -14,7 +14,7 @@ public interface VHSRepository extends JpaRepository<VHS, Long> {
 	public List<VHS> getAllVHS();
 	
 	@Query(value="SELECT * FROM vhs WHERE vhs_id = :id", nativeQuery=true)
-	public VHS getVHSById(@Param("id") int id);
+	public VHS getVHSById(@Param("id") long id);
 	
 	@Query(value="SELECT * FROM vhs WHERE title = :title", nativeQuery=true)
 	public List<VHS> getVHSByTitle(@Param("title") String title);
